requirements.txt
# core web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# async database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ml frameworks
torch==2.1.2
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.9.4

# data processing
pandas==2.2.0

# background jobs (optional for later)
apscheduler==3.10.4

# testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# utilities
python-dotenv==1.0.0
